backend/requirement.txt
celery==4.4
djangorestframework
requests
django-cors-headers
serial
